Rename hdf exporter to hdf legacy exporter, to make room for a standardized hdf exporter
